scripts/requirements.txt
requests>=2.28.0
python-telegram-bot>=20.0
feedparser>=6.0.10
beautifulsoup4>=4.11.0
pandas>=1.5.0
python-dotenv>=1.0.0
yfinance>=0.2.36
